RAC3: Refactor: Extract World and Webworld classes
