+++ worlds/rac3/options/armor_upgrade_options.py
@@ -1,5 +1,5 @@
 from Options import Range
-from worlds.rac3 import RAC3OPTION
+from worlds.rac3.constants.options import RAC3OPTION
 
 
 class ArmorUpgrade(Range):
