+++ worlds/rac3/options/filler_weight_options.py
@@ -1,6 +1,7 @@
 from Options import ItemDict
-from worlds.rac3 import RAC3ITEM, RAC3OPTION
 from worlds.rac3.constants.data.item import filler_data
+from worlds.rac3.constants.items import RAC3ITEM
+from worlds.rac3.constants.options import RAC3OPTION
 
 
 class FillerWeight(ItemDict):
