+++ worlds/rac3/options/one_hp_options.py
@@ -1,5 +1,5 @@
 from Options import OptionCounter
-from worlds.rac3 import RAC3OPTION
+from worlds.rac3.constants.options import RAC3OPTION
 from worlds.rac3.constants.player_type import ONE_HP_CHALLENGE_CHARACTERS
 
 
@@ -12,12 +12,11 @@ class OneHpChallenge(OptionCounter):
     Vehicle: Vehicles have only 1 HP.
     Qwark: Qwark has only 1 HP.
     Giant Clank: Giant Clank has only 1 HP. This one is impossible as terror of talos will one hit you immediately
-    If you feel like this is too hard for you, you can always toggle the challenge for a character using /one_hp [character name]
+    If you feel like this is too hard for you, you can always toggle the challenge for a character using /one_hp [
+    character name]
     """
     min = 0
     max = 1
     display_name = RAC3OPTION.ONE_HP_CHALLENGE
     default = {name: 0 for name in ONE_HP_CHALLENGE_CHARACTERS}
     valid_keys = ONE_HP_CHALLENGE_CHARACTERS
-
-
