+++ worlds/rac3/options/sewer_limitation_options.py
@@ -1,5 +1,6 @@
 from Options import Range
-from worlds.rac3 import RAC3OPTION
+from worlds.rac3.constants.options import RAC3OPTION
+
 
 class SewerLimitation(Range):
     """
