+++ worlds/rac3/universal_tracker.py
@@ -6,7 +6,7 @@ from worlds.rac3.constants.options import RAC3OPTION
 from worlds.rac3.constants.region import RAC3REGION
 
 if TYPE_CHECKING:
-    from worlds.rac3 import RaC3World
+    from worlds.rac3.world import RaC3World
 
 
 def setup_options_from_slot_data(world: "RaC3World") -> None:
